lara: tidy structure
