+++ meson.build
@@ -121,15 +121,14 @@ dll_sources = [
   'src/game/inventory/ring.c',
   'src/game/inventory/vars.c',
   'src/game/items.c',
+  'src/game/lara/cheat.c',
+  'src/game/lara/col.c',
   'src/game/lara/common.c',
-  'src/game/lara/lara_cheat.c',
-  'src/game/lara/lara_col.c',
-  'src/game/lara/lara_control.c',
-  'src/game/lara/lara_draw.c',
-  'src/game/lara/lara_look.c',
-  'src/game/lara/lara_misc.c',
-  'src/game/lara/lara_state.c',
+  'src/game/lara/control.c',
+  'src/game/lara/draw.c',
+  'src/game/lara/look.c',
   'src/game/lara/misc.c',
+  'src/game/lara/state.c',
   'src/game/level.c',
   'src/game/los.c',
   'src/game/lot.c',
