+++ src/decomp/decomp.c
@@ -12,8 +12,8 @@
 #include "game/inventory/backpack.h"
 #include "game/inventory/common.h"
 #include "game/items.h"
-#include "game/lara/lara_control.h"
-#include "game/lara/lara_draw.h"
+#include "game/lara/control.h"
+#include "game/lara/draw.h"
 #include "game/lot.h"
 #include "game/math.h"
 #include "game/music.h"
