+++ src/game/console/cmd/teleport.c
@@ -3,7 +3,7 @@
 #include "game/game_string.h"
 #include "game/gameflow/gameflow_new.h"
 #include "game/items.h"
-#include "game/lara/lara_cheat.h"
+#include "game/lara/cheat.h"
 #include "game/objects/names.h"
 #include "game/objects/vars.h"
 #include "game/random.h"
