+++ src/game/gun/gun_rifle.c
@@ -5,7 +5,7 @@
 #include "game/gun/gun_misc.h"
 #include "game/input.h"
 #include "game/items.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/misc.h"
 #include "game/math.h"
 #include "game/random.h"
 #include "game/sound.h"
