+++ src/game/lara/col.c
@@ -1,10 +1,10 @@
-#include "game/lara/lara_col.h"
+#include "game/lara/col.h"
 
 #include "game/collide.h"
 #include "game/input.h"
 #include "game/items.h"
-#include "game/lara/lara_control.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/control.h"
+#include "game/lara/misc.h"
 #include "game/math.h"
 #include "game/room.h"
 #include "game/sound.h"
