+++ src/game/lara/control.c
@@ -1,13 +1,13 @@
-#include "game/lara/lara_control.h"
+#include "game/lara/control.h"
 
 #include "game/creature.h"
 #include "game/gun/gun.h"
 #include "game/input.h"
 #include "game/inventory/backpack.h"
 #include "game/items.h"
-#include "game/lara/lara_cheat.h"
-#include "game/lara/lara_look.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/cheat.h"
+#include "game/lara/look.h"
+#include "game/lara/misc.h"
 #include "game/math.h"
 #include "game/music.h"
 #include "game/room.h"
