+++ src/game/lara/draw.h
