+++ src/game/lara/state.c
@@ -1,10 +1,10 @@
-#include "game/lara/lara_state.h"
+#include "game/lara/state.h"
 
 #include "game/input.h"
 #include "game/inventory/backpack.h"
-#include "game/lara/lara_control.h"
-#include "game/lara/lara_look.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/control.h"
+#include "game/lara/look.h"
+#include "game/lara/misc.h"
 #include "game/music.h"
 #include "game/output.h"
 #include "game/room.h"
