+++ src/game/objects/common.c
@@ -1,7 +1,7 @@
 #include "game/objects/common.h"
 
 #include "game/items.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/misc.h"
 #include "global/funcs.h"
 #include "global/vars.h"
