+++ src/game/objects/effects/flame.c
@@ -2,8 +2,8 @@
 
 #include "game/collide.h"
 #include "game/effects.h"
-#include "game/lara/lara_control.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/control.h"
+#include "game/lara/misc.h"
 #include "game/room.h"
 #include "game/sound.h"
 #include "global/funcs.h"
