+++ src/game/objects/vehicles/boat.c
@@ -4,7 +4,7 @@
 #include "game/effects.h"
 #include "game/input.h"
 #include "game/items.h"
-#include "game/lara/lara_look.h"
+#include "game/lara/look.h"
 #include "game/math.h"
 #include "game/objects/common.h"
 #include "game/random.h"
