+++ src/game/room.c
@@ -3,7 +3,7 @@
 #include "game/box.h"
 #include "game/camera.h"
 #include "game/items.h"
-#include "game/lara/lara_misc.h"
+#include "game/lara/misc.h"
 #include "game/lot.h"
 #include "game/math.h"
 #include "game/music.h"
