+++ src/game/room_draw.c
@@ -1,7 +1,7 @@
 #include "game/room_draw.h"
 
 #include "game/effects.h"
-#include "game/lara/lara_draw.h"
+#include "game/lara/draw.h"
 #include "game/matrix.h"
 #include "game/output.h"
 #include "global/funcs.h"
