+++ src/inject_exec.c
@@ -22,13 +22,13 @@
 #include "game/inventory/common.h"
 #include "game/inventory/ring.h"
 #include "game/items.h"
-#include "game/lara/lara_cheat.h"
-#include "game/lara/lara_col.h"
-#include "game/lara/lara_control.h"
-#include "game/lara/lara_draw.h"
-#include "game/lara/lara_look.h"
-#include "game/lara/lara_misc.h"
-#include "game/lara/lara_state.h"
+#include "game/lara/cheat.h"
+#include "game/lara/col.h"
+#include "game/lara/control.h"
+#include "game/lara/draw.h"
+#include "game/lara/look.h"
+#include "game/lara/misc.h"
+#include "game/lara/state.h"
 #include "game/level.h"
 #include "game/los.h"
 #include "game/lot.h"
