+++ src/specific/s_input.c
@@ -5,7 +5,7 @@
 #include "game/gameflow/gameflow_new.h"
 #include "game/input.h"
 #include "game/inventory/backpack.h"
-#include "game/lara/lara_control.h"
+#include "game/lara/control.h"
 #include "global/const.h"
 #include "global/funcs.h"
 #include "global/vars.h"
